+++ server/src/lsp/hover_handler.rs
@@ -574,6 +574,7 @@ impl Backend {
                     event_targets: Some(&self.scanner_data.event_targets),
                     characters: Some(&self.scanner_data.characters),
                     achievements: Some(achievements),
+                    variables: Some(&self.scanner_data.variables),
                     in_random_list: false,
                     state_targeted: false,
                 };